+++ R/oauth-flow-auth-code.R
@@ -29,8 +29,8 @@
 #' @inheritParams req_perform
 #' @param cache_disk Should the access token be cached on disk? This reduces
 #'   the number of times that you need to re-authenticate at the cost of
-#'   storing access credentials on disk. Cached tokens are encrypted and
-#'   automatically deleted 30 days after creation.
+#'   storing access credentials on disk. Cached tokens are encrypted,
+#'   automatically deleted 30 days after creation, and stored in [cache_path()].
 #' @param cache_key If you want to cache multiple tokens per app, use this
 #'   key to disambiguate them.
 #' @returns A modified HTTP [request].